task: filter products by category

+++ spa/src/services/product.js
@@ -2,22 +2,22 @@ import api from "./api";
 
 const productService = {
   create: (data) => {
-    return api.post("products", data);
+    return api.post("admin/products", data);
   },
   findAll: () => {
-    return api.get("products");
+    return api.get("admin/products");
   },
 
   findById: (id) => {
-    return api.get(`products/${id}`);
+    return api.get(`admin/products/${id}`);
   },
 
   productRemove: (id) => {
-    return api.delete(`products/${id}`);
+    return api.delete(`admin/products/${id}`);
   },
 
   productUpdate: (data) => {
-    return api.put(`products/${data.id}`, data);
+    return api.put(`admin/products/${data.id}`, data);
   },
 };
